+++ lib/ex_check/check.ex
@@ -5,6 +5,7 @@ defmodule ExCheck.Check do
   alias ExCheck.Check.Pipeline
   alias ExCheck.Command
   alias ExCheck.Config
+  alias ExCheck.Manifest
   alias ExCheck.Printer
 
   def run(opts) do
@@ -14,8 +15,10 @@ defmodule ExCheck.Check do
       |> Keyword.delete(:config)
       |> Config.load()
 
-    opts = Keyword.merge(config_opts, opts)
-    opts = convert_failed_to_only_with_manifest(opts)
+    opts =
+      config_opts
+      |> Keyword.merge(opts)
+      |> Manifest.convert_failed_to_only()
 
     compile_and_run_tools(tools, opts)
   end
@@ -33,7 +36,7 @@ defmodule ExCheck.Check do
 
     reprint_errors(failed_results)
     print_summary(all_results, total_duration, opts)
-    save_manifest(failed_results, opts)
+    Manifest.save(all_results, opts)
     maybe_set_exit_status(failed_results)
   end
 
@@ -278,47 +281,6 @@ defmodule ExCheck.Check do
     sec_str = if sec < 10, do: "0#{sec}", else: "#{sec}"
 
     "#{min}:#{sec_str}"
-  end
-
-  defp convert_failed_to_only_with_manifest(opts) do
-    with true <- Keyword.get(opts, :failed),
-         manifest_path = get_manifest_path(opts),
-         true <- File.exists?(manifest_path) do
-      only =
-        manifest_path
-        |> File.read!()
-        |> String.split("\n")
-        |> Enum.filter(fn
-          "" -> false
-          _check -> true
-        end)
-        |> Enum.map(&{:only, String.to_atom(&1)})
-        |> case do
-          [] -> [{:only, "-"}]
-          only -> only
-        end
-
-      opts ++ only
-    else
-      _ -> opts
-    end
-  end
-
-  defp save_manifest(failed_tools, opts) do
-    failed_tool_names = Enum.map(failed_tools, fn {:error, {name, _, _}, _} -> name end)
-    manifest_content = Enum.join(failed_tool_names, "\n") <> "\n"
-    manifest_path = get_manifest_path(opts)
-
-    File.write!(manifest_path, manifest_content)
-  end
-
-  @escape Enum.map(' [~#%&*{}\\:<>?/+|"]', &<<&1::utf8>>)
-
-  defp get_manifest_path(opts) do
-    Keyword.get_lazy(opts, :failure_manifest, fn ->
-      app_id = File.cwd!() |> String.replace(@escape, "_") |> String.replace(~r/^_+/, "")
-      Path.join([System.tmp_dir!(), "ex_check-failure_manifest-#{app_id}.txt"])
-    end)
   end
 
   defp maybe_set_exit_status(failed_tools) do